GET, POSTリクエストを抽象化した
