--- Frontend/src/js/main/actions/api/api_utils.js
@@ -1,67 +1,0 @@
-import Cookies from 'js-cookie';
-
-/**
- * API認証用のトークンを取得する
- * @returns {*}
- */
-export function genAuthToken() {
-    const token = Cookies.get('access_token');
-    return {Authorization: 'Bearer ' + token};
-}
-
-/**
- * Fetch APIでGETするときの設定を返す
- * @returns {{method: string, mode: string, headers: {Authorization: string}}}
- */
-export function getOptions() {
-    const auth = genAuthToken();
-    return {
-        method: 'GET',
-        mode: 'cors',
-        headers: {
-            ...auth
-        },
-    };
-}
-
-/**
- * Fetch APIでPOSTするときの設定を返す
- * @returns {{method: string, mode: string, headers: {Accept: string, Content-Type: string, Authorization: string}}}
- */
-export function postOptions() {
-    const auth = genAuthToken();
-    return {
-        method: 'POST',
-        mode: 'cors',
-        headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json',
-            ...auth,
-        },
-    };
-}
-
-/**
- * サーバーから返ってきたレスポンスのステータスコードを判定し処理する。
- * 失敗している場合はPromise.reject()で返す。
- * 成功している場合はJSONファイルをPromiseで返す。
- * @param {Response} response Fetch APIのResponseオブジェクト
- * @returns {Promise}
- */
-export function checkStatus(response) {
-    if (!response.ok) {
-        return Promise.reject(new Error(response.status));
-    } else {
-        return response.json();
-    }
-}
-
-/**
- * ログインページに移動する
- */
-export function redirectToLogin() {
-    if(!location){
-        return;
-    }
-    location.href = '/';
-}
